refactor: 값 전달을 위한 dto 객체 추가 (#1)
